pyproject.toml
[tool.poetry]
name = "qr-code-studio"
version = "0.1.0"
description = ""
authors = ["TFP <thefoxpaw@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
dearpygui = "^1.6.2"
qrcode = "^7.3.1"
Pillow = "^9.2.0"
pyzbar = "^0.1.9"
qrcode-term = "^0.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
